+++ packages/@atjson/offset-annotations/src/annotations/threads-embed.ts
@@ -1,0 +1,28 @@
+import { BlockAnnotation } from "@atjson/document";
+
+export class ThreadsEmbed extends BlockAnnotation<{
+  /**
+   * The URL to the Threads post
+   */
+  url: string;
+
+  /**
+   * Content of the Threads post, which allows to store
+   * text of the post content if that is wanted.
+   */
+  content?: string;
+
+  /**
+   * Layout information, used to indicate mutually
+   * exclusive layouts, for example sizes, floats, etc.
+   */
+  layout?: string;
+
+  /**
+   * A named identifier used to quickly jump to this item
+   */
+  anchorName?: string;
+}> {
+  static type = "threads-embed";
+  static vendorPrefix = "offset";
+}