+++ packages/@atjson/offset-annotations/src/utils/convert-html-tables.ts
@@ -7,7 +7,12 @@ import {
   TextAnnotation,
 } from "@atjson/document";
 
-import OffsetSource, { DataSet, Table, ColumnType } from "../index";
+import OffsetSource, {
+  DataSet,
+  Table,
+  ColumnType,
+  TextAlignment,
+} from "../index";
 
 const generateId = hexoid(8);
 
@@ -75,8 +80,13 @@ function extractAlignment(tableCell: Annotation<{ style: string }>) {
       /(text-align: ?(?<alignment>left|right|center))/
     );
 
-    if (match?.groups?.alignment) {
-      return match.groups.alignment as "left" | "right" | "center";
+    switch (match?.groups?.alignment) {
+      case "left":
+        return TextAlignment.Start;
+      case "right":
+        return TextAlignment.End;
+      case "center":
+        return TextAlignment.Center;
     }
   }
 
@@ -149,7 +159,7 @@ export function convertHTMLTablesToDataSet(
 
         let alignment = extractAlignment(headCell);
         if (alignment) {
-          columnConfig.textAlign = alignment;
+          columnConfig.textAlignment = alignment;
         }
 
         columnConfigs.push(columnConfig);
@@ -196,7 +206,7 @@ export function convertHTMLTablesToDataSet(
 
             let alignment = extractAlignment(bodyCell);
             if (alignment) {
-              columnConfig.textAlign = alignment;
+              columnConfig.textAlignment = alignment;
             }
 
             columnConfigs[index] = columnConfig;