+++ packages/@atjson/renderer-plain-text/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.24.43](https://github.com/CondeNast/atjson/compare/@atjson/renderer-plain-text@0.24.42...@atjson/renderer-plain-text@0.24.43) (2024-06-12)
+
+**Note:** Version bump only for package @atjson/renderer-plain-text
+
+## [0.24.42](https://github.com/CondeNast/atjson/compare/@atjson/renderer-plain-text@0.24.41...@atjson/renderer-plain-text@0.24.42) (2024-06-10)
+
+**Note:** Version bump only for package @atjson/renderer-plain-text
+
+## [0.24.41](https://github.com/CondeNast/atjson/compare/@atjson/renderer-plain-text@0.24.40...@atjson/renderer-plain-text@0.24.41) (2024-05-30)
+
+**Note:** Version bump only for package @atjson/renderer-plain-text
+
+## [0.24.40](https://github.com/CondeNast/atjson/compare/@atjson/renderer-plain-text@0.24.39...@atjson/renderer-plain-text@0.24.40) (2024-05-30)
+
+**Note:** Version bump only for package @atjson/renderer-plain-text
+
+## [0.24.39](https://github.com/CondeNast/atjson/compare/@atjson/renderer-plain-text@0.24.38...@atjson/renderer-plain-text@0.24.39) (2024-05-30)
+
+**Note:** Version bump only for package @atjson/renderer-plain-text
+
 ## [0.24.38](https://github.com/CondeNast/atjson/compare/@atjson/renderer-plain-text@0.24.37...@atjson/renderer-plain-text@0.24.38) (2024-05-16)
 
 **Note:** Version bump only for package @atjson/renderer-plain-text