+++ packages/@atjson/source-commonmark/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.27.4](https://github.com/CondeNast/atjson/compare/@atjson/source-commonmark@0.27.3...@atjson/source-commonmark@0.27.4) (2024-06-12)
+
+**Note:** Version bump only for package @atjson/source-commonmark
+
+## [0.27.3](https://github.com/CondeNast/atjson/compare/@atjson/source-commonmark@0.27.2...@atjson/source-commonmark@0.27.3) (2024-06-10)
+
+**Note:** Version bump only for package @atjson/source-commonmark
+
+## [0.27.2](https://github.com/CondeNast/atjson/compare/@atjson/source-commonmark@0.27.1...@atjson/source-commonmark@0.27.2) (2024-05-30)
+
+**Note:** Version bump only for package @atjson/source-commonmark
+
+## [0.27.1](https://github.com/CondeNast/atjson/compare/@atjson/source-commonmark@0.27.0...@atjson/source-commonmark@0.27.1) (2024-05-30)
+
+**Note:** Version bump only for package @atjson/source-commonmark
+
+# [0.27.0](https://github.com/CondeNast/atjson/compare/@atjson/source-commonmark@0.26.3...@atjson/source-commonmark@0.27.0) (2024-05-30)
+
+### Features
+
+- attribute standardization ([#1685](https://github.com/CondeNast/atjson/issues/1685)) ([36f709f](https://github.com/CondeNast/atjson/commit/36f709f3bcf0e68dd09afd17c69e35b76ea7c7a8))
+
 ## [0.26.3](https://github.com/CondeNast/atjson/compare/@atjson/source-commonmark@0.26.2...@atjson/source-commonmark@0.26.3) (2024-05-16)
 
 ### Bug Fixes