+++ packages/@atjson/source-mobiledoc/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.23.31](https://github.com/CondeNast/atjson/compare/@atjson/source-mobiledoc@0.23.30...@atjson/source-mobiledoc@0.23.31) (2024-06-12)
+
+**Note:** Version bump only for package @atjson/source-mobiledoc
+
+## [0.23.30](https://github.com/CondeNast/atjson/compare/@atjson/source-mobiledoc@0.23.29...@atjson/source-mobiledoc@0.23.30) (2024-06-10)
+
+**Note:** Version bump only for package @atjson/source-mobiledoc
+
+## [0.23.29](https://github.com/CondeNast/atjson/compare/@atjson/source-mobiledoc@0.23.28...@atjson/source-mobiledoc@0.23.29) (2024-05-30)
+
+**Note:** Version bump only for package @atjson/source-mobiledoc
+
+## [0.23.28](https://github.com/CondeNast/atjson/compare/@atjson/source-mobiledoc@0.23.27...@atjson/source-mobiledoc@0.23.28) (2024-05-30)
+
+**Note:** Version bump only for package @atjson/source-mobiledoc
+
+## [0.23.27](https://github.com/CondeNast/atjson/compare/@atjson/source-mobiledoc@0.23.26...@atjson/source-mobiledoc@0.23.27) (2024-05-30)
+
+**Note:** Version bump only for package @atjson/source-mobiledoc
+
 ## [0.23.26](https://github.com/CondeNast/atjson/compare/@atjson/source-mobiledoc@0.23.25...@atjson/source-mobiledoc@0.23.26) (2024-05-16)
 
 **Note:** Version bump only for package @atjson/source-mobiledoc